move i18n translation logic into shared bundle
