+++ services/QuillLMS/client/app/bundles/Diagnostic/styles/style.scss
@@ -54,7 +54,9 @@ body {
   }
 }
 
-
+.activity-container {
+  background-color: $quill-grey-2;
+}
 
 .has-bottom-margin {
   margin-bottom: 20px;
