+++ services/QuillLMS/client/app/bundles/Diagnostic/styles/variables.scss
@@ -17,7 +17,7 @@ $purple: #847bb9;
 $red: #ed6c63;
 $turquoise: #1fc8db;
 $yellow: #fce473;
-$background: #eeeeee;
+$background: $quill-grey-2;
 $disabled: #bcbcbc;
 
 $primary: $green;
