+++ services/QuillLMS/client/app/bundles/Shared/libs/index.ts
@@ -43,4 +43,4 @@ export {
   sortByLevenshteinAndOptimal,
   extractConceptResultsFromResponse
 } from './responseTools'
-export { getlanguageOptions, hasTranslationFlag, showTranslations } from './translations'
+export { getlanguageOptions, hasTranslationFlag, showTranslations } from './translations/helpers'
