+++ services/QuillLMS/client/app/bundles/Shared/libs/translations/modules/arabic/translation.json
