+++ services/QuillLMS/client/app/bundles/Shared/libs/translations/modules/chinese/translation.json
