+++ services/QuillLMS/client/app/bundles/Shared/libs/translations/modules/dari/translation.json
