+++ services/QuillLMS/client/app/bundles/Shared/libs/translations/modules/french/translation.json
