+++ services/QuillLMS/client/app/bundles/Shared/libs/translations/modules/german/translation.json
