+++ services/QuillLMS/client/app/bundles/Shared/libs/translations/modules/languagePageInfo.js
@@ -1,4 +1,4 @@
-import {ENGLISH, CHINESE, HINDI, SPANISH, FRENCH, ARABIC, RUSSIAN, PORTUGUESE, URDU, GERMAN, JAPANESE, KOREAN, VIETNAMESE, THAI, UKRAINIAN, TAGALOG, DARI } from "../../../Shared/utils/languageList"
+import {ENGLISH, CHINESE, HINDI, SPANISH, FRENCH, ARABIC, RUSSIAN, PORTUGUESE, URDU, GERMAN, JAPANESE, KOREAN, VIETNAMESE, THAI, UKRAINIAN, TAGALOG, DARI } from "../../../utils/languageList"
 
 const ELL_STARTER_PRE_ARCHIVED = '-LyFRZvbHAmooTTIIVE2'
 const ELL_STARTER_PRE = '873840f0-2ecf-4dcb-bb5b-9acfe3c0b969'
