+++ services/QuillLMS/client/app/bundles/Shared/libs/translations/modules/portuguese/translation.json
