+++ services/QuillLMS/client/app/bundles/Shared/libs/translations/modules/spanish/translation.json
