+++ services/QuillLMS/client/app/bundles/Shared/styles/activity_container.scss
@@ -53,7 +53,7 @@
       }
     }
     .quill-button {
-      border: none;
+      border: none !important;
       font-weight: 600;
     }
   }
