+++ services/QuillLMS/client/app/bundles/Shared/styles/button-archived.scss
@@ -86,4 +86,5 @@
 
 #teacher-preview-menu-button {
   cursor: pointer;
+  border: none;
 }
